MVC Fonctionnel

+++ Pratique MVC Graph/ConsoleApp1/Models/Path.cs
@@ -11,14 +11,14 @@ namespace LOTRMaps.Models
     {
         public Location From { get; set; }
         public Location To { get; set; }
-        public float Distance { get; set; } 
-        public float Time { get; set; }
+        public double Distance { get; set; } 
+        public double Time { get; set; }
         public override Model FromDataRow(DataRow row)
         {
             this.From = new Location() { Short = row["from"].ToString() };
             this.To = new Location() { Short = row["to"].ToString() };
-            this.Distance = (float)row["Distance"];
-            this.Time = (float)row["Time"];
+            this.Distance = (double)row["Distance"];
+            this.Time = (double)row["Time"];
             return this;
         }
     }
